requirements.txt
mysql-connector-python
google-cloud-vision